refactor cli.js
